// front-end/src/shared/interfaces/cartProducts/ICartProducts.ts
import type { IProduct } from "../products/IProducts";

export interface ICartProduct {
  id: number;
  productId: number;
  quantity: number;
  cartId: number;
  product?: IProduct;
  createdAt: Date;
  updatedAt: Date;
}

export interface ICreateCartProduct {
  id?: number;
  quantity: number;
  cartId: number;
  productId: number;
}

export interface IUpdateCartProduct {
  id?: number;
  quantity?: number;
  cartId?: number;
  productId?: number;
}